backend/requirements.txt
opencv-python
numpy
ultralytics
flask-cors
requests
python-dotenv
